refactor(staking): inline page UI into server component
Move all UI from _components/staking.tsx into app/[locale]/staking/page.tsx
as a single async server component, swapping the client useTranslation hook
for next-intl's getTranslations. Add "use client" to StakingHierarchy since
it owns onClick handlers that previously inherited client-ness from the
removed client wrapper.
